+++ test/Fixture/CallLikes/NoNamedArgumentRule/InvokableClass.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Ergebnis\PHPStan\Rules\Test\Fixture\CallLikes\NoNamedArgumentRule;
+
+final class InvokableClass
+{
+    public function __invoke($bar): void
+    {
+    }
+}