+++ CHANGELOG.md
@@ -11,6 +11,7 @@ For a full diff see [`2.10.3...main`][2.10.3...main].
 ### Fixed
 
 - Adjusted `Methods\NoNamedArgumentRule` to handle static calls on variable expressions ([#947]), by [@localheinz]
+- Adjusted `Methods\NoNamedArgumentRule` to handle calls on invokables ([#948]), by [@localheinz]
 
 ## [`2.10.3`][2.10.3]
 
@@ -689,6 +690,7 @@ For a full diff see [`362c7ea...0.1.0`][362c7ea...0.1.0].
 [#943]: https://github.com/ergebnis/phpstan-rules/pull/943
 [#944]: https://github.com/ergebnis/phpstan-rules/pull/944
 [#947]: https://github.com/ergebnis/phpstan-rules/pull/947
+[#948]: https://github.com/ergebnis/phpstan-rules/pull/948
 
 [@cosmastech]: https://github.com/cosmastech
 [@enumag]: https://github.com/enumag